segment race tests
