+++ nodebuilder/node_test.go
@@ -1,3 +1,5 @@
+//go:build !race
+
 package nodebuilder
 
 import (
